Moved /about and /team commands to a new feature
