+++ version.py
@@ -1,1 +1,1 @@
-VERSION = "0.9.7"
+VERSION = "0.9.8"
